[datajs][lg]: Moved src code outside
