--- datajs/src/browser-utils/index.js
@@ -1,7 +1,0 @@
-if (typeof window !== 'undefined') module.exports = { ...require('./utils') }
-else
-  module.exports = {
-    toNodeStream: (reader, size, returnChunk) => {},
-    isFileFromBrowser: (file) => {},
-    webToNodeStream: (reader) => {},
-  }
