--- datajs/src/index.js
@@ -1,14 +1,0 @@
-export {
-  open,
-  parsePath,
-  parseDatasetIdentifier,
-  isDataset,
-  isUrl,
-} from './data'
-export { File, FileInterface } from './file-base'
-export { FileLocal } from './file-local'
-export { FileRemote } from './file-remote'
-export { FileInline } from './file-inline'
-export { Dataset } from './dataset'
-export { csvParser } from './parser/csv'
-export { xlsxParser } from './parser/xlsx'
