+++ package.json
@@ -11,12 +11,12 @@
     "test:browser": "yarn run build:browser && karma start karma.conf.js --browsers=ChromeHeadless",
     "test:node": "nyc mocha --require @babel/register datajs/test/*",
     "build:browser": "rm -rf ./dist/browser && webpack --mode production",
-    "build:node": "rm -rf ./dist/node && babel ./datajs/src -d ./dist/node --no-comments",
+    "build:node": "rm -rf ./dist/node && babel ./datajs/ -d ./dist/node --no-comments",
     "build": "npm run build:browser && npm run build:node",
     "build:watch": "webpack --watch",
     "coveralls": "cat ./coverage/lcov.info | ./node_modules/.bin/coveralls",
     "coverage": "nyc report --reporter=text-lcov | coveralls && nyc report --reporter=lcov",
-    "lint": "eslint ./datajs/src"
+    "lint": "eslint ./datajs"
   },
   "repository": {
     "type": "git",
